requirements.txt
playwright>=1.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
openpyxl>=3.1.2
reportlab>=4.1.0
beautifulsoup4>=4.12.3
lxml>=4.9.4
pytest>=7.4.0
pytest-asyncio>=0.21.1